wrap networks to allow additional networks to be added
